LPS-128276 Move to the proper package
