+++ modules/apps/headless/headless-admin-content/headless-admin-content-impl/src/main/java/com/liferay/headless/admin/content/internal/vulcan/graphql/contributor/StructuredContentGraphQLContributor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.headless.admin.content.internal.graphql.extension;
+package com.liferay.headless.admin.content.internal.vulcan.graphql.contributor;
 
 import com.liferay.portal.vulcan.graphql.contributor.GraphQLContributor;
